porządki, dodanie możliwości wpisania miasta zamiast podania lokalizacji, wstęp do nawigacji z favourites do weather view
