+++ WeatherApp.xcodeproj/xcuserdata/hackentosh.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "4E827A0E-49F6-48B5-88B9-E693AE268C47"
+   type = "1"
+   version = "2.0">
+</Bucket>
