+++ WeatherApp/WeatherAppApp.swift
@@ -1,7 +1,7 @@
 import SwiftUI
 
 @main
-struct WeatherAppApp: App {
+struct WeatherAppApp: App {		
     var body: some Scene {
         WindowGroup {
             ContentView()
